# .fape/runs/20260810-195134-workspace-save-load/docs/05-openapi.yaml
openapi: 3.1.0
info:
  title: Workspace Save/Load API
  version: 1.0.0
  description: >
    Endpoints added by run 20260810-195134-workspace-save-load. Only the
    changed/added surface is documented here — the rest of this app's API
    (generation jobs, assets) is unchanged and out of scope for this file.
servers:
  - url: /api
paths:
...
  /workspace-saves/{id}/objects/{objectId}/file:
    get:
      operationId: getWorkspaceSaveObjectFile
      summary: Stream a saved upload-sourced object's persisted GLB bytes (FR-6, AC-6)
      parameters:
        - $ref: "#/components/parameters/SaveId"
        - name: objectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: GLB binary stream
          content:
            model/gltf-binary:
              schema:
                type: string
                format: binary
        "404":
          description: >
            Save not found, object not found, object is not an upload-sourced
            object, or the stored file is missing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    SaveId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            type: string
...